+++ algafood-api/src/main/java/com/algaworks/algafood/infraestructure/repository/CozinhaRepositoryImpl.java
@@ -3,6 +3,7 @@ package com.algaworks.algafood.infraestructure.repository;
 import com.algaworks.algafood.domain.model.Cozinha;
 import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,7 +30,7 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 
     @Transactional
     @Override
-    public Cozinha adicionar(Cozinha cozinha) {
+    public Cozinha salvar(Cozinha cozinha) {
         return entityManager.merge(cozinha);
     }
 
@@ -37,6 +38,11 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
     @Override
     public void remover(Long id) {
         Cozinha cozinha = porID(id);
+
+        if(cozinha == null){
+            throw new EmptyResultDataAccessException(1);
+        }
+
         entityManager.remove(cozinha);
     }
 }
